requirements.txt
pydantic==2.9.2
plotly==5.24.1
flask
pyyaml